refactor: moved epoch_train and epoch_evaluate to utils
