+++ src/utils/__init__.py
@@ -1,3 +1,4 @@
 from utils.utils import *
 from utils.data import *
 from utils.random import *
+from utils.epoch import epoch_evaluate, epoch_train
